Сhange the names of queries. Edit checkFavoriteGenre query. Add JWT. (#3)
* Сhange the names of queries. Edit checkFavoriteGenre query.

* Add JWT

+++ ApolloServer/resolvers/Genre.js
@@ -1,30 +1,33 @@
 const { getRepository } = require("typeorm");
+const { ApolloError } = require("apollo-server-express");
 const { FavoriteGenre } = require("../../build/entities/FavoriteGenre");
 const { genresUrlPart } = require("../utils");
 
 const resolvers = {
   Query: {
-    getGenres: async (parent, { language }, { getResponse }) => {
+    genres: async (parent, { language }, { getResponse, user }) => {
+      if (!user) throw new ApolloError("Authorization failed.", "UNAUTORIZED");
       const urlPart = genresUrlPart(language);
       const genres = await getResponse(urlPart);
       return genres ? genres.genres : [];
     },
-    checkFavoriteGenre: async (parent, { userId, genreId }) => {
-      const genre = await getRepository(FavoriteGenre).findOne({
-        genreId: genreId,
-        userId: userId,
+    favoriteGenres: async (parent, args, { user }) => {
+      if (!user) throw new ApolloError("Authorization failed.", "UNAUTORIZED");
+      const favoriteGenres = await getRepository(FavoriteGenre).find({
+        userId: user.id,
       });
-      return Boolean(genre);
+      return favoriteGenres.map((current) => current.genreId);
     },
   },
   Mutation: {
     setFavoriteGenres: async (
       parent,
-      { userId, genreId },
-      { generateAnswer }
+      { genreId },
+      { generateAnswer, user }
     ) => {
+      if (!user) throw new ApolloError("Authorization failed.", "UNAUTORIZED");
       const genreRepository = getRepository(FavoriteGenre);
-      const params = { userId: userId, genreId: genreId };
+      const params = { userId: user.id, genreId };
       const favoriteGenre = await genreRepository.findOne(params);
       if (favoriteGenre) genreRepository.delete(params);
       else genreRepository.save(params);
